Add action_abstraction input + TODOs
Adding a vector slice input containing the bet sizes (exact in Chips, not ratios) that we want to translate to. Currently TODO to actually use it properly.

TBD: Might want to instead make this take the action_abstraction as values as ratios *normalized to the pot size*, not 100% sure which will approach will be more user friendly in practice