Add+use TranslatedAction struct return type
I think this is basically the MVP for psuedoharmonic action translation here. Going to probably add a couple more test cases and then wrap things up here.

(Side note: still need to double check whether the annotations on the struct are correct/expected; cursory search hasn't shown me a definitive answer yet on the best way to structure things when creating a struct just to serve as the return type like this in Rust)