Tweak name since there's a gotcha here
Almost forgot - other important thing this test is exercising: specifically, scale invariance. If we were to naively impelment translate_actions where we plugged in the numbers without normalizing to pot size we'd get the wrong results.

(Specifically for

f(A,B, x) = ((B - x) * (1 + A)) / ((B - A) * (1 + x))

if we don't normalize properly and just directly plug in A=25, B=75, x=50 it equals 13/51 => ~0.25. Which is WAY off from the 5/12 => ~0.4167 that is actually correct.