Darn. ... well, that kinda sucks
Checked the logs to confirm aaaaand we're hitting the bug again (oof). Seems that  using f32::MIN doesn't actually get the assertion to work properly? ... I guess there might actaully be more of a difference than we saw in the logs. Regardless, the difference SHOULD be a lot bigger, so easy fix.

Probably something unrelated to the swaps that somehow got touched/reverted etc. Not really sure what exactly a the moment unfortunately.

Unfortuantely gotta leave as is for now - will have to come back and poker around more later.