WIP - quick and dirty vibe code start
(Not fully vibe coded - the llm started throwing in a BUNCH of unneeded stuff + made a few mistakes so I had to edit a decent chunk by hand. But, it actually overall got pretty darn close!

Annoyingly the test case for the unaccelerated algorithm actually hit my assert block (so temporarily commenting it out). ... which is strange because the RMS is *definitely* decreasing when I run it in reality...?
```
04:38:51 [INFO] clustering kmeans (unaccelerated)turn
04:39:34 [DEBUG] abstraction cluster RMS error   0.038611025
04:40:42 [DEBUG] abstraction cluster RMS error   0.036001097
```
... quite odd!)