Fixed tests: use Arbitrary crate + deterministic psuedo-rng
I re-ran them like 10 times in a row and the only error was when
```
---- clustering::emd::tests::is_sinkhorn_emd_zero stdout ----
thread 'clustering::emd::tests::is_sinkhorn_emd_zero' panicked at src/clustering/emd.rs:125:9:
consider decreasing temp or tolerance
0.0102291275 0.01
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```
flaked once. Which isn't even this part of the code :)

____

Somehow I missed that histogram ALREADY has a .random(), so need to hardcode the individual abstraction values myself.

Switching over to it (+ seeded rng) immediately fixed the issue with elkan and original not matching, now they are decreasing in lockstep :)
And, I was able to deal with the other issues by simply creating a (lot) more points for the centers, making the proportion of centers to points more realistic and so avoids running into the 'no points assigned to this center' edge case