... actually we might have a problem after all
This actually fails! Despite (AIUI) the resulting histogram inputs in theory being the same.
Am wondering if it's because the resulting inputs are different, vs because there was something getting improperly shared between them in the old approach using clone... (would need to dig into how rust clones work a bit more to know for sure)