Remove unused _rng + actually fix the issues with the test here
... the _rng wasn't even being used. None of this was deterministic at all (DANGIT CLAUDE). ... This will teach me to look more carefully next time :)

____

TLDR the actual solution WAS to use .clone() to guarantee the inputs going in are the same. Since we're not using anything 'exotic' here, in Rust there's no way mutations to one can't affect the other, so that wasn't a real concern at all :)

Now, technically this mean we _could_ get super unlucky with how the random point histograms end up lining up with the starting center histograms. But in practice? This is perfectly fine - ran it 10+ times and no errors or flakes. (... after changing it to be 10x less strict on the difference between them; otherwise I did actually get a rare flake.)