Added spinner ... but broke the main progress bar styling again
WIP / possibly revert and redo.

Working on a multiprogress that allows the main bar to go, while also doing a spinner right below it. I got the second part (spinner) going, but now the original progress bar isn't properly rendering (and we're getting a bunch of empty lines written out at the bottom for no reason)