Finally works perfectly!
Turns out that all we had to do was get it ticking on a constant schedule to get the progress updates to show up. (I think was probably because it was just not flushing stuff? Probably could do it another way, though this is totally fine as-is)

Still need to go fix one other progress bar but this one is golden now :D

...  ... I guess if we're being REALLY pedantic, maybe we should be adding + removing the spinner at the 'longer' steps doing the distance calculations. Possibly with a message saying 'step X in progress' or something. also similarly would be nice to make the bars in the tests pretty. But gonna call it a night here / tackle that tomorrow.